requirements.txt
nicegui>=1.4
pymongo>=4.7
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0
py-mini-racer>=0.6
